Projet/web/requirements.txt
Flask>=2.0
flask-cors
requests
gunicorn
paho-mqtt
